create projectile spawn module and ProjectileSpawnPlugin
